--- Example_Cases/case9_en_net_.txt
@@ -1,23 +1,0 @@
-3 6 
-K 
-0 1 13.6257 
-0 2 10.6537 
-1 0 50.3299 
-1 2 34.873 
-2 0 83.6725 
-2 1 74.1486 
-P 
-0 -1.8145 -1.8145 
-1 34.3647 34.3647 
-2 35.2895 35.2895 
-alfa 
-0 1.0575 
-1 3.9063 
-2 8.3056 
-gamma 
-0 1 -0.18754 
-0 2 -0.16937 
-1 0 -0.18754 
-1 2 -0.19358 
-2 0 -0.16937 
-2 1 -0.19358
