--- Example_Cases/case9_sm_net_.txt
@@ -1,167 +1,0 @@
-9 72 
-K 
-0 1 5.6118 
-0 2 4.6371 
-0 3 3.9655 
-0 4 11.759 
-0 5 8.621 
-0 6 0.99923 
-0 7 3.6081 
-0 8 1.1293 
-1 0 20.7286 
-1 2 17.022 
-1 3 2.3826 
-1 4 21.8484 
-1 5 8.6849 
-1 6 11.3862 
-1 7 28.6993 
-1 8 4.1454 
-2 0 36.419 
-2 1 36.1929 
-2 3 4.1861 
-2 4 19.8304 
-2 5 28.8439 
-2 6 6.4445 
-2 7 42.4729 
-2 8 20.8067 
-3 0 937.4389 
-3 1 152.4868 
-3 2 126.0015 
-3 4 319.5202 
-3 5 234.2552 
-3 6 27.1517 
-3 7 98.0421 
-3 8 30.6851 
-4 0 55.5964 
-4 1 27.966 
-4 2 11.9379 
-4 3 6.3904 
-4 5 14.9896 
-4 6 4.9796 
-4 7 14.1075 
-4 8 2.9072 
-5 0 56.6116 
-5 1 15.4399 
-5 2 24.1167 
-5 3 6.5071 
-5 4 20.8189 
-5 6 2.7492 
-5 7 13.8531 
-5 8 5.8731 
-6 0 236.2189 
-6 1 728.7144 
-6 2 193.9787 
-6 3 27.1517 
-6 4 248.9795 
-6 5 98.9716 
-6 7 327.051 
-6 8 47.2396 
-7 0 21.3241 
-7 1 45.9189 
-7 2 31.9609 
-7 3 2.4511 
-7 4 17.6344 
-7 5 12.4678 
-7 6 8.1763 
-7 8 7.7834 
-8 0 266.9598 
-8 1 265.3029 
-8 2 626.2824 
-8 3 30.6851 
-8 4 145.3619 
-8 5 211.4328 
-8 6 47.2396 
-8 7 311.3369 
-P 
-0 4.1875  4.1875  
-1 44.4477 44.4477 
-2 47.3353 47.3353 
-3 -4.7625 -4.7625 
-4 -51.363 -51.363 
-5 -51.8721 -51.8721 
-6 -7.2232 -7.2232 
-7 -50.2443 -50.2443 
-8 -10.5222 -10.5222 
-alfa 
-0 1.0575 
-1 3.9063 
-2 8.3056 
-3 250 
-4 5 
-5 6.9444 
-6 250 
-7 6.25 
-8 250 
-gamma 
-0 1 0.10544 
-0 2 0.114 
-0 3 -0.044213 
-0 4 -0.010895 
-0 5 0.0016932 
-0 6 0.10544 
-0 7 0.12663 
-0 8 0.114 
-1 0 0.10544 
-1 2 0.012426 
-1 3 0.10544 
-1 4 0.053796 
-1 5 0.10081 
-1 6 -0.055697 
-1 7 -0.024334 
-1 8 0.012426 
-2 0 0.114 
-2 1 0.012426 
-2 3 0.114 
-2 4 0.1109 
-2 5 0.031669 
-2 6 0.012426 
-2 7 -0.02692 
-2 8 -0.069045 
-3 0 -0.044213 
-3 1 0.10544 
-3 2 0.114 
-3 4 -0.010895 
-3 5 0.0016932 
-3 6 0.10544 
-3 7 0.12663 
-3 8 0.114 
-4 0 -0.010895 
-4 1 0.053796 
-4 2 0.1109 
-4 3 -0.010895 
-4 5 0.040686 
-4 6 0.053796 
-4 7 0.087904 
-4 8 0.1109 
-5 0 0.0016932 
-5 1 0.10081 
-5 2 0.031669 
-5 3 0.0016932 
-5 4 0.040686 
-5 6 0.10081 
-5 7 0.076524 
-5 8 0.031669 
-6 0 0.10544 
-6 1 -0.055697 
-6 2 0.012426 
-6 3 0.10544 
-6 4 0.053796 
-6 5 0.10081 
-6 7 -0.024334 
-6 8 0.012426 
-7 0 0.12663 
-7 1 -0.024334 
-7 2 -0.02692 
-7 3 0.12663 
-7 4 0.087904 
-7 5 0.076524 
-7 6 -0.024334 
-7 8 -0.02692 
-8 0 0.114 
-8 1 0.012426 
-8 2 -0.069045 
-8 3 0.114 
-8 4 0.1109 
-8 5 0.031669 
-8 6 0.012426 
-8 7 -0.02692
