--- Example_Cases/case9_sp_net_.txt
@@ -1,65 +1,0 @@
-12 24 
-K 
-0 3 144.1157 
-1 4 264.6431 
-2 5 360.0537 
-3 0 136.2758 
-3 6 139.6464 
-4 1 67.7486 
-4 9 126.8395 
-5 2 43.3505 
-5 11 136.1493 
-6 3 139.6464 
-6 7 89.9733 
-6 8 83.7147 
-7 6 89.9733 
-7 9 46.9105 
-8 6 83.7147 
-8 11 45.1922 
-9 4 126.8395 
-9 7 46.9105 
-9 10 108.3718 
-10 9 108.3718 
-10 11 77.9053 
-11 5 136.1493 
-11 8 45.1922 
-11 10 77.9053 
-P 
-0 5.7124 5.7124 
-1 48.0075 48.0075 
-2 53.2296 53.2296 
-3 0 0
-4 0 0
-5 0 0
-6 -26.2398 -26.2398 
-7 -28.5046 -28.5046 
-8 -31.7149 -31.7149 
-9 -22.2511 -22.2511 
-10 -29.111 -29.111 
-11 -19.5835 -19.5835 
-alfa 
-0 1.0575 
-1 3.9063 
-2 8.3056 
-3 0 
-4 0 
-5 0 
-6 0 
-7 0 
-8 0 
-9 0 
-10 0 
-11 0 
-gamma 
-6 7 0.11711 
-6 8 0.18272 
-7 6 0.11711 
-7 9 0.1962 
-8 6 0.18272 
-8 11 0.22551 
-9 7 0.1962 
-9 10 0.11751 
-10 9 0.11751 
-10 11 0.11751 
-11 8 0.22551 
-11 10 0.11751
