--- Example_Cases/initstate_case9_en_net_.txt
@@ -1,6 +1,0 @@
-0 
-0.16197 
-0.081415 
-0 
-0 
-0
