--- Example_Cases/initstate_case9_sm_net_.txt
@@ -1,18 +1,0 @@
-0 
-0.16197 
-0.081415 
--0.03869 
--0.069618 
--0.064357 
-0.064921 
-0.012698 
-0.034326 
-0 
-0 
-0 
-0 
-0 
-0 
-0 
-0 
-0
