--- Example_Cases/initstate_case9_sp_net_.txt
@@ -1,24 +1,0 @@
-0 
-0 
-0 
-0 
-0 
-0 
-0 
-0 
-0 
-0 
-0 
-0 
-0 
-0 
-0 
-0 
-0 
-0 
-0 
-0 
-0 
-0 
-0 
-0
