+++ Initial_States/initstate_case9_kinit_1_sm_.txt
@@ -1,0 +1,18 @@
+0.019464149339554385
+0.022985115407101346
+0.023421186333128882
+0.0190102468699187
+0.015176818664141917
+0.015183909040284122
+0.019010443840725297
+0.015170014217701831
+0.019009364423924292
+0.0
+0.0
+0.0
+0.0
+0.0
+0.0
+0.0
+0.0
+0.0
