+++ Initial_States/initstate_case9_kinit_1_sp_.txt
@@ -1,15 +1,15 @@
 0.0
-0.17493160792532358
-0.0896259125041262
+0.16875136718264663
+0.08327093684252854
 0.0
 0.0
 0.0
--0.03885700783994458
--0.07265955125389889
--0.06552664873044457
-0.07208242033012385
-0.015886157425508488
-0.0398781445401137
+-0.042003860318715135
+-0.07592066315404244
+-0.07011448942395995
+0.06630715604148234
+0.010847998939381095
+0.03360808951711517
 0.0
 0.0
 0.0
