+++ Initial_States/initstate_qr_sm_net_0_deltd_0.5_0_.txt
@@ -1,0 +1,250 @@
+1.4519682787776012 
+4.439696374657064 
+0.4734256119587119 
+0.5717076946950604 
+4.629567526939376 
+4.900004389645601 
+2.5229689101728754 
+1.432521117226895 
+4.7568884589943075 
+2.0774560974894465 
+4.579893731891776 
+3.186664408106323 
+4.03975476402794 
+0.7025845099278748 
+2.9574090725382103 
+4.0672588336207065 
+1.4429285419735782 
+5.81192110462521 
+2.0043253865061024 
+3.9589982410397346 
+0.173792584296539 
+0.08355751953447069 
+0.3826690130378353 
+5.807863336454088 
+4.158928809864555 
+1.8520508954215604 
+3.9975249074997476 
+1.4196958976511336 
+1.1829044802045212 
+5.415265240913988 
+1.6486613178354064 
+0.8512022827498696 
+5.161566203520092 
+4.927581284856753 
+5.830808001745327 
+2.368079734117787 
+1.151780231668717 
+1.482224419861756 
+2.8990651898346482 
+3.846056655227667 
+5.4895169057848685 
+5.593812533159598 
+4.474066957058045 
+2.686539287982123 
+4.789626059120224 
+3.7718606847487948 
+2.1333579555710487 
+0.04498113817698846 
+4.591645590820918 
+5.510244263392546 
+3.0184674039718757 
+2.4831995998880614 
+4.335578945873551 
+0.7955678985836871 
+4.946691208035304 
+1.5703798083326017 
+3.5387062390302666 
+3.6704158725619958 
+6.214859374313977 
+0.9178435000277271 
+0.8226118228550156 
+3.2639652856598595 
+3.363263323072339 
+3.89734953886735 
+0.7660248325882238 
+5.683741994386648 
+4.8496415875536245 
+5.344315364415751 
+1.7809428207346396 
+0.34550464028208716 
+1.5967494936113475 
+6.0087297274631135 
+5.661424311519339 
+4.417878315365133 
+0.6412388934772163 
+6.1823629689485875 
+4.528630401611673 
+4.0405227555905405 
+0.43785004407690414 
+1.0748990910014955 
+1.0804535419567063 
+1.0028130804900024 
+6.136826483924362 
+5.559549453840578 
+1.239389411284137 
+5.3273187992195306 
+0.04486777570051433 
+2.1490483265491775 
+2.833317308126198 
+0.7906309324750002 
+0.2617407844211008 
+4.914834490545675 
+4.342422853546157 
+3.172133884459896 
+5.991551793441099 
+3.759938872717337 
+3.291784873323209 
+4.683258921367465 
+4.558482629871482 
+5.956539563455581 
+5.182610376975792 
+3.079649254070512 
+5.31352972375043 
+0.8449414027564589 
+5.1395180002970395 
+3.3113392701994355 
+0.025234644293842767 
+2.9195399495089016 
+6.033469290941849 
+2.4124630687359656 
+4.458663313776654 
+5.499707295888415 
+1.9178658988760313 
+1.6918109473856255 
+4.021984640838074 
+1.8893408511140275 
+2.2234061113949903 
+3.798262506131562 
+2.962480104911532 
+3.600008714379714 
+1.2741166864132172 
+6.244145374241454 
+2.541120001520745 
+4.424584576447998 
+4.073126581786135 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0
