+++ Initial_States/initstate_rd_sm_net_0_deltd_0.5_0_.txt
@@ -1,0 +1,250 @@
+3.9857555948903425 
+2.838275199545593 
+1.3768546490754632 
+2.1870798949384676 
+2.936983493028853 
+3.970684097072173 
+1.8231965613859833 
+3.6639369036895544 
+5.541545927240249 
+0.5471867951325321 
+4.008865590444922 
+6.222783768839545 
+3.5281633099910468 
+3.5895449987129786 
+4.737700798660331 
+0.3233588844057752 
+3.3825290234329226 
+4.28895561732552 
+3.1171895453049197 
+2.790150788120864 
+0.4496701557535781 
+3.4214822904682625 
+3.6556491244991247 
+5.90159333452928 
+4.776561346791034 
+0.9510069339707306 
+3.5659604942800165 
+1.6392772862747205 
+2.8766685371761875 
+4.639955518472941 
+1.430153727938554 
+3.6619065422318213 
+2.337962031975482 
+1.3912943932720214 
+3.8613278333606114 
+4.64378596038205 
+1.4344088039617724 
+4.2331483179876805 
+0.2191160762219619 
+4.297219564819469 
+2.909428920017233 
+3.972475962699858 
+2.6106864952904973 
+0.38902050961803797 
+6.070626140748828 
+1.3998061136291309 
+4.815344213659925 
+1.6332603143249484 
+4.684895417558306 
+5.594171532387794 
+4.981692548972928 
+2.018036257685427 
+1.388528894737817 
+1.2250437037222084 
+0.3400696566409053 
+1.493841471206802 
+0.0851546008042917 
+3.490859170993875 
+0.34866359954206844 
+3.9859399308658867 
+1.5039566571246479 
+3.976684036349387 
+3.591455493081709 
+5.2964880539169075 
+2.1809946113004974 
+4.07692883981468 
+0.7119152850353733 
+4.425981679317986 
+0.8887197342413894 
+2.2003648571304533 
+3.052144384379128 
+4.955173535222677 
+3.0933810806988378 
+5.545602282158175 
+3.6237308755252853 
+3.4695210800792524 
+0.7191612596863 
+4.307767274058731 
+4.052681344724628 
+2.2473543217875243 
+2.248590123354314 
+5.506379394524489 
+2.0985641259276324 
+4.078755614615251 
+0.1888508554192409 
+1.6218987538249099 
+0.10344998910518163 
+3.615850943077049 
+1.9912959996011985 
+0.10198942833330919 
+1.1773852621421186 
+5.470018932891789 
+5.073671449697282 
+5.0117104699342745 
+5.53682724934572 
+2.751510578283279 
+4.253801872122624 
+3.4421204421058627 
+2.942963553739959 
+4.072399072914055 
+2.509296734061368 
+3.9943345577612703 
+6.259648868050388 
+3.675473632470787 
+3.380214995163732 
+4.552139367671599 
+3.5695989433939443 
+3.7249111572740286 
+1.1060853417352434 
+5.748766611465287 
+3.906316566534949 
+2.2643565071306737 
+4.6797998968240355 
+2.8144344935247507 
+1.7762717060205548 
+5.639411951282791 
+5.262633537628452 
+3.5367902469562966 
+5.90866734786548 
+0.7339840650814226 
+5.760212110921876 
+5.631335444345272 
+2.5323186007349774 
+3.207706806305254 
+5.752182611513049 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0
