+++ Initial_States/initstate_sw_sm_net_0_deltd_0.5_0_.txt
@@ -1,0 +1,250 @@
+2.709505073876606 
+4.103493265261258 
+1.7230961288581077 
+1.011281644149809 
+5.124586625655576 
+1.1765125959221012 
+5.930056709081869 
+0.014858657908426732 
+2.8673744224151076 
+4.142183617844064 
+4.8890800454185115 
+3.897389506532989 
+0.7019941113373491 
+3.059470995936115 
+2.3466419841151227 
+5.632030842186524 
+6.2124620883260055 
+3.458958046986397 
+2.408788726914936 
+4.258516479116724 
+2.7955026558111165 
+0.8496319070370028 
+2.675274882347098 
+1.5150780983196557 
+6.18055626458259 
+1.898341737787661 
+1.2134480950011857 
+5.904751516591011 
+1.6022236315718386 
+3.4963708267609657 
+1.516020964854236 
+3.5428632887081584 
+5.483034463521181 
+1.7444894962039401 
+3.1941992250396867 
+6.154738797979109 
+4.26949170246129 
+0.7525596276757481 
+2.425697026473869 
+1.822776636036773 
+5.0680551707823005 
+1.7327864183107602 
+1.0715950854262626 
+0.6509219996816532 
+0.6371223482607978 
+5.604078469651604 
+1.6431870937931223 
+5.566284796113946 
+2.0411961877246276 
+1.68706603199771 
+3.100149299852491 
+6.247464240456727 
+4.025334297732835 
+2.4636986982223794 
+6.2112819462896995 
+5.054292714962879 
+5.095258062816408 
+4.570789542007933 
+3.6663021520349863 
+0.3811326475401035 
+5.761891865439801 
+2.385201747575486 
+2.1209681472659927 
+4.168465907642441 
+1.8237727200602907 
+2.154471257533303 
+3.5989534857692322 
+4.947098083928634 
+1.285739017434351 
+5.472832917981312 
+0.09881665338218282 
+1.1705263716479377 
+1.257578996561275 
+0.3695129646415597 
+3.6253563472320987 
+4.8841509328042605 
+1.2380430768375097 
+0.6925053473865 
+0.3485354266595422 
+3.7599154417896803 
+4.535793831449262 
+0.3843254356840317 
+1.294134845731072 
+1.7531125653876973 
+0.9072694963609449 
+2.7810258621890362 
+4.565916375916277 
+6.1794046910394345 
+2.6775231828850248 
+4.571019179579493 
+4.726799988027702 
+1.9435564488999362 
+1.6524892705757361 
+0.09169739379542308 
+3.1160988491538393 
+1.590948106838425 
+1.719490860408296 
+3.6350592987879615 
+0.185347758828275 
+3.269693450912524 
+2.6668251038313886 
+4.753949742292291 
+2.739360153294266 
+3.572979678108029 
+1.579315162472954 
+3.6805296909760488 
+2.3344349041654655 
+0.08104989002768474 
+2.6260913804823485 
+0.2605843657892905 
+1.2851843391744617 
+5.940024455468411 
+1.875870622153957 
+5.899917507531946 
+3.3536531605159197 
+3.4951215066168966 
+6.056051699678721 
+4.534924386382784 
+1.3235221596602575 
+4.489292148347509 
+1.0035072833223517 
+0.5384913623772379 
+0.7120911808012814 
+1.7162070575601205 
+2.729508158292231 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0 
+0.0
