+++ Networks/case9_kinit_1_sm_.txt
@@ -1,0 +1,95 @@
+9 72 0 
+K 
+0 1 5.15592388260155 
+0 2 4.316765065079815 
+0 3 3.6376387383831306 
+0 4 11.06773898902378 
+0 5 8.162350754501457 
+0 6 0.8956125496453597 
+0 7 3.711731365085154 
+0 8 1.0198265829059967 
+1 0 19.04469384135947 
+1 2 15.985839417275656 
+1 3 2.2072787401810334 
+1 4 21.119167337571724 
+1 5 8.25848069192889 
+1 6 10.665264557694378 
+1 7 30.61121690129256 
+1 8 3.77662062725732 
+2 0 33.90309838487934 
+2 1 33.98982467460605 
+2 3 3.929366831232202 
+2 4 18.980401571136404 
+2 5 28.308234301285847 
+2 6 5.904220898517718 
+2 7 45.57793569070508 
+2 8 19.72109182280621 
+3 0 859.9377977537721 
+3 1 141.26583937158614 
+3 2 118.27394162008927 
+3 4 303.24214899797454 
+3 5 223.63816007270879 
+3 6 24.538659114870324 
+3 7 101.69677806531703 
+3 8 27.941967633350647 
+4 0 52.328269940104434 
+4 1 27.03253419209181 
+4 2 11.426201745824116 
+4 3 6.06484297995949 
+4 5 14.654550149427672 
+4 6 4.695700988304469 
+4 7 15.161016535641641 
+4 8 2.699415906671049 
+5 0 53.59943662122622 
+5 1 14.68174345231803 
+5 2 23.66882923524178 
+5 3 6.2121711131307995 
+5 4 20.353541874205092 
+5 6 2.550300195652788 
+5 7 14.934172595008286 
+5 8 5.591710662140441 
+6 0 211.72280673616302 
+6 1 682.5769316924401 
+6 2 177.71704904538333 
+6 3 24.538659114870324 
+6 4 234.78504941522345 
+6 5 91.81080704350038 
+6 7 340.30963238044325 
+6 8 41.98527557550543 
+7 0 21.936332367653257 
+7 1 48.97794704206809 
+7 2 34.29739660725558 
+7 3 2.5424194516329255 
+7 4 18.951270669552045 
+7 5 13.440755335507458 
+7 6 8.50774080951108 
+7 8 8.102687141233725 
+8 0 241.08700419897755 
+8 1 241.7037201444685 
+8 2 593.6048638664668 
+8 3 27.941967633350643 
+8 4 134.97079533355247 
+8 5 201.30158383705592 
+8 6 41.98527557550543 
+8 7 324.1074856493489 
+P 
+0 1.0 5.342302228185688 5.342302228185688 
+1 1.0 48.007462737669016 48.007462737669016 
+2 1.0 53.22964296780048 53.22964296780048 
+3 0.0 -0.0 -0.0 
+4 0.0 -47.12388980384689 -47.12388980384689 
+5 0.0 -47.12388980384689 -47.12388980384689 
+6 0.0 -0.0 -0.0 
+7 0.0 -47.12388980384689 -47.12388980384689 
+8 0.0 -0.0 -0.0 
+Alfa 
+0 1.0575296108291032 
+1 3.90625 
+2 8.305647840531561 
+3 250.0 
+4 5.0 
+5 6.944444444444445 
+6 250.0 
+7 6.25 
+8 250.0 
+Gamma
