+++ Networks/case9_kinit_1_sp_.txt
@@ -1,46 +1,46 @@
-12 24 0 
+12 24 12 
 K 
-0 3 1.0 
-1 4 1.0 
-2 5 1.0
-3 0 1.0 
-3 6 1.0
-4 1 1.0
-4 9 1.0 
-5 2 1.0
-5 11 1.0 
-6 3 1.0
-6 7 1.0
-6 8 1.0
-7 6 1.0
-7 9 1.0 
-8 6 1.0
-8 11 1.0 
-9 4 1.0 
-9 7 1.0 
-9 10 1.0 
-10 9 1.0 
-10 11 1.0
-11 5 1.0 
-11 8 1.0 
-11 10 1.0
+0 3 134.92557115253808 
+1 4 254.13587518077813 
+2 5 350.097707534246 
+3 0 127.58562008184002 
+3 6 129.19889485488312 
+4 1 65.0587840462792 
+4 9 120.17695679417535 
+5 2 42.15176398712322 
+5 11 129.10021433252842 
+6 3 129.19889485488312 
+6 7 83.2666605347286 
+6 8 77.59213549742132 
+7 6 83.26666053472859 
+7 9 43.8183283416462 
+8 6 77.59213549742132 
+8 11 42.31094477990471 
+9 4 120.17695679417535 
+9 7 43.8183283416462 
+9 10 102.113622728844 
+10 9 102.11362272884398 
+10 11 73.46475051730037 
+11 5 129.10021433252842 
+11 8 42.3109447799047 
+11 10 73.46475051730039 
 P 
-0 1.0 10.0 10.0 
-1 1.0 2.5 2.5 
-2 1.0 2.5 2.5 
+0 1.0 5.661245644793789 5.661245644793789 
+1 1.0 48.007462737669016 48.007462737669016 
+2 1.0 53.22964296780048 53.22964296780048 
 3 0.0 -0.0 -0.0 
 4 0.0 -0.0 -0.0 
 5 0.0 -0.0 -0.0 
-6 0.0 -0.0 -0.0 
-7 0.0 -5.0 -5.0 
-8 0.0 -3.0 -3.0 
-9 0.0 -0.0 -0.0 
-10 0.0 -7.0 -7.0 
-11 0.0 -0.0 -0.0 
+6 0.0 -24.293238259141805 -24.293238259141805 
+7 0.0 -27.07555867974564 -27.07555867974564 
+8 0.0 -29.91782575280201 -29.91782575280201 
+9 0.0 -20.986107702093296 -20.986107702093296 
+10 0.0 -27.845999727347003 -27.845999727347003 
+11 0.0 -18.499534115007574 -18.499534115007574 
 Alfa 
-0 1.0 
-1 1.0 
-2 1.0 
+0 1.0575296108291032 
+1 3.90625 
+2 8.305647840531561 
 3 0.0 
 4 0.0 
 5 0.0 
@@ -51,3 +51,15 @@ Alfa
 10 0.0 
 11 0.0 
 Gamma 
+6 7 0.11710874455836429 
+6 8 0.18272156075289758 
+7 6 0.11710874455836429 
+7 9 0.19620081727201644 
+8 6 0.18272156075289758 
+8 11 0.22550963517142386 
+9 7 0.19620081727201644 
+9 10 0.11751164542424553 
+10 9 0.11751164542424553 
+10 11 0.11751164542136538 
+11 8 0.22550963517142386 
+11 10 0.11751164542136538
