+++ Networks/sw_sm_net_0_deltd_0.0_.txt
@@ -1,0 +1,664 @@
+110 440 
+K 
+0 1 1 
+0 2 1 
+0 108 1 
+0 109 1 
+1 0 1 
+1 2 1 
+1 3 1 
+1 109 1 
+2 0 1 
+2 1 1 
+2 3 1 
+2 4 1 
+3 1 1 
+3 2 1 
+3 4 1 
+3 5 1 
+4 2 1 
+4 3 1 
+4 5 1 
+4 6 1 
+4 67 1 
+5 3 1 
+5 4 1 
+5 45 1 
+5 84 1 
+6 4 1 
+6 7 1 
+6 8 1 
+7 6 1 
+7 8 1 
+7 9 1 
+8 6 1 
+8 7 1 
+8 9 1 
+8 10 1 
+9 7 1 
+9 8 1 
+9 10 1 
+9 11 1 
+10 8 1 
+10 9 1 
+10 11 1 
+10 12 1 
+11 9 1 
+11 10 1 
+11 12 1 
+11 65 1 
+12 10 1 
+12 11 1 
+12 13 1 
+12 14 1 
+12 50 1 
+13 12 1 
+13 15 1 
+13 20 1 
+14 12 1 
+14 15 1 
+14 89 1 
+15 13 1 
+15 14 1 
+15 16 1 
+15 17 1 
+16 15 1 
+16 17 1 
+16 18 1 
+17 15 1 
+17 16 1 
+17 18 1 
+17 19 1 
+18 16 1 
+18 17 1 
+18 19 1 
+18 20 1 
+19 17 1 
+19 18 1 
+19 21 1 
+19 105 1 
+20 13 1 
+20 18 1 
+20 21 1 
+20 22 1 
+21 19 1 
+21 20 1 
+21 22 1 
+21 23 1 
+22 20 1 
+22 21 1 
+22 23 1 
+22 24 1 
+23 21 1 
+23 22 1 
+23 24 1 
+23 25 1 
+24 22 1 
+24 23 1 
+24 25 1 
+24 26 1 
+25 23 1 
+25 24 1 
+25 26 1 
+25 27 1 
+25 38 1 
+26 24 1 
+26 25 1 
+26 27 1 
+26 28 1 
+27 25 1 
+27 26 1 
+27 28 1 
+27 29 1 
+28 26 1 
+28 27 1 
+28 29 1 
+28 30 1 
+29 27 1 
+29 28 1 
+29 31 1 
+29 83 1 
+30 28 1 
+30 31 1 
+30 32 1 
+31 29 1 
+31 30 1 
+31 32 1 
+31 33 1 
+32 30 1 
+32 31 1 
+32 33 1 
+32 34 1 
+33 31 1 
+33 32 1 
+33 34 1 
+33 35 1 
+34 32 1 
+34 33 1 
+34 35 1 
+34 36 1 
+35 33 1 
+35 34 1 
+35 36 1 
+35 37 1 
+36 34 1 
+36 35 1 
+36 37 1 
+36 38 1 
+37 35 1 
+37 36 1 
+37 38 1 
+37 39 1 
+38 25 1 
+38 36 1 
+38 37 1 
+38 40 1 
+39 37 1 
+39 40 1 
+39 41 1 
+40 38 1 
+40 39 1 
+40 41 1 
+40 42 1 
+41 39 1 
+41 40 1 
+41 42 1 
+41 43 1 
+41 102 1 
+42 40 1 
+42 41 1 
+42 43 1 
+42 44 1 
+43 41 1 
+43 42 1 
+43 44 1 
+43 45 1 
+44 42 1 
+44 43 1 
+44 45 1 
+44 46 1 
+45 5 1 
+45 43 1 
+45 44 1 
+45 46 1 
+45 47 1 
+46 44 1 
+46 45 1 
+46 47 1 
+46 48 1 
+46 53 1 
+47 45 1 
+47 46 1 
+47 48 1 
+47 49 1 
+48 46 1 
+48 47 1 
+48 49 1 
+48 50 1 
+48 73 1 
+49 47 1 
+49 48 1 
+49 50 1 
+49 51 1 
+50 12 1 
+50 48 1 
+50 49 1 
+50 52 1 
+51 49 1 
+51 52 1 
+51 53 1 
+52 50 1 
+52 51 1 
+52 53 1 
+52 54 1 
+53 46 1 
+53 51 1 
+53 52 1 
+53 55 1 
+54 52 1 
+54 55 1 
+54 56 1 
+54 88 1 
+55 53 1 
+55 54 1 
+55 56 1 
+55 57 1 
+56 54 1 
+56 55 1 
+56 57 1 
+56 58 1 
+57 55 1 
+57 56 1 
+57 58 1 
+57 59 1 
+58 56 1 
+58 57 1 
+58 59 1 
+58 60 1 
+59 57 1 
+59 58 1 
+59 60 1 
+59 61 1 
+60 58 1 
+60 59 1 
+60 61 1 
+60 62 1 
+61 59 1 
+61 60 1 
+61 62 1 
+61 63 1 
+62 60 1 
+62 61 1 
+62 63 1 
+62 64 1 
+63 61 1 
+63 62 1 
+63 64 1 
+63 65 1 
+64 62 1 
+64 63 1 
+64 65 1 
+64 66 1 
+65 11 1 
+65 63 1 
+65 64 1 
+65 66 1 
+65 67 1 
+66 64 1 
+66 65 1 
+66 67 1 
+66 95 1 
+66 104 1 
+67 4 1 
+67 65 1 
+67 66 1 
+67 68 1 
+68 67 1 
+68 69 1 
+68 89 1 
+69 68 1 
+69 70 1 
+69 71 1 
+70 69 1 
+70 71 1 
+70 72 1 
+71 69 1 
+71 70 1 
+71 72 1 
+71 73 1 
+71 85 1 
+72 70 1 
+72 71 1 
+72 73 1 
+72 74 1 
+73 48 1 
+73 71 1 
+73 72 1 
+73 74 1 
+74 72 1 
+74 73 1 
+74 75 1 
+74 76 1 
+75 74 1 
+75 76 1 
+75 77 1 
+76 74 1 
+76 75 1 
+76 77 1 
+76 78 1 
+77 75 1 
+77 76 1 
+77 78 1 
+77 79 1 
+78 76 1 
+78 77 1 
+78 79 1 
+78 80 1 
+79 77 1 
+79 78 1 
+79 80 1 
+79 81 1 
+80 78 1 
+80 79 1 
+80 81 1 
+80 82 1 
+81 79 1 
+81 80 1 
+81 83 1 
+81 92 1 
+82 80 1 
+82 83 1 
+82 84 1 
+83 29 1 
+83 81 1 
+83 82 1 
+83 84 1 
+83 85 1 
+84 5 1 
+84 82 1 
+84 83 1 
+84 85 1 
+84 86 1 
+85 71 1 
+85 83 1 
+85 84 1 
+85 87 1 
+86 84 1 
+86 87 1 
+86 88 1 
+87 85 1 
+87 86 1 
+87 88 1 
+87 89 1 
+88 54 1 
+88 86 1 
+88 87 1 
+88 90 1 
+89 14 1 
+89 68 1 
+89 87 1 
+89 90 1 
+89 91 1 
+90 88 1 
+90 89 1 
+90 91 1 
+90 92 1 
+91 89 1 
+91 90 1 
+91 92 1 
+91 93 1 
+92 81 1 
+92 90 1 
+92 91 1 
+92 93 1 
+92 94 1 
+93 91 1 
+93 92 1 
+93 94 1 
+93 95 1 
+94 92 1 
+94 93 1 
+94 95 1 
+94 96 1 
+95 66 1 
+95 93 1 
+95 94 1 
+95 97 1 
+96 94 1 
+96 97 1 
+96 98 1 
+97 95 1 
+97 96 1 
+97 98 1 
+97 99 1 
+98 96 1 
+98 97 1 
+98 99 1 
+98 100 1 
+99 97 1 
+99 98 1 
+99 100 1 
+99 101 1 
+100 98 1 
+100 99 1 
+100 101 1 
+100 102 1 
+101 99 1 
+101 100 1 
+101 102 1 
+101 103 1 
+102 41 1 
+102 100 1 
+102 101 1 
+102 103 1 
+103 101 1 
+103 102 1 
+103 104 1 
+103 105 1 
+104 66 1 
+104 103 1 
+104 105 1 
+104 106 1 
+105 19 1 
+105 103 1 
+105 104 1 
+105 106 1 
+105 107 1 
+106 104 1 
+106 105 1 
+106 107 1 
+106 108 1 
+107 105 1 
+107 106 1 
+107 108 1 
+107 109 1 
+108 0 1 
+108 106 1 
+108 107 1 
+108 109 1 
+109 0 1 
+109 1 1 
+109 107 1 
+109 108 1 
+P 
+0 0.0 -1.0 -1.0 
+1 0.0 -1.0 -1.0 
+2 0.0 -1.0 -1.0 
+3 1.0 10.0 10.0 
+4 0.0 -1.0 -1.0 
+5 0.0 -1.0 -1.0 
+6 0.0 -1.0 -1.0 
+7 0.0 -1.0 -1.0 
+8 0.0 -1.0 -1.0 
+9 0.0 -1.0 -1.0 
+10 0.0 -1.0 -1.0 
+11 0.0 -1.0 -1.0 
+12 0.0 -1.0 -1.0 
+13 0.0 -1.0 -1.0 
+14 0.0 -1.0 -1.0 
+15 0.0 -1.0 -1.0 
+16 0.0 -1.0 -1.0 
+17 0.0 -1.0 -1.0 
+18 0.0 -1.0 -1.0 
+19 0.0 -1.0 -1.0 
+20 0.0 -1.0 -1.0 
+21 0.0 -1.0 -1.0 
+22 0.0 -1.0 -1.0 
+23 0.0 -1.0 -1.0 
+24 0.0 -1.0 -1.0 
+25 0.0 -1.0 -1.0 
+26 0.0 -1.0 -1.0 
+27 0.0 -1.0 -1.0 
+28 0.0 -1.0 -1.0 
+29 0.0 -1.0 -1.0 
+30 1.0 10.0 10.0 
+31 0.0 -1.0 -1.0 
+32 0.0 -1.0 -1.0 
+33 0.0 -1.0 -1.0 
+34 0.0 -1.0 -1.0 
+35 1.0 10.0 10.0 
+36 0.0 -1.0 -1.0 
+37 0.0 -1.0 -1.0 
+38 0.0 -1.0 -1.0 
+39 0.0 -1.0 -1.0 
+40 0.0 -1.0 -1.0 
+41 0.0 -1.0 -1.0 
+42 0.0 -1.0 -1.0 
+43 0.0 -1.0 -1.0 
+44 0.0 -1.0 -1.0 
+45 0.0 -1.0 -1.0 
+46 0.0 -1.0 -1.0 
+47 0.0 -1.0 -1.0 
+48 0.0 -1.0 -1.0 
+49 0.0 -1.0 -1.0 
+50 0.0 -1.0 -1.0 
+51 0.0 -1.0 -1.0 
+52 0.0 -1.0 -1.0 
+53 0.0 -1.0 -1.0 
+54 0.0 -1.0 -1.0 
+55 0.0 -1.0 -1.0 
+56 0.0 -1.0 -1.0 
+57 0.0 -1.0 -1.0 
+58 0.0 -1.0 -1.0 
+59 0.0 -1.0 -1.0 
+60 0.0 -1.0 -1.0 
+61 0.0 -1.0 -1.0 
+62 1.0 10.0 10.0 
+63 0.0 -1.0 -1.0 
+64 0.0 -1.0 -1.0 
+65 0.0 -1.0 -1.0 
+66 0.0 -1.0 -1.0 
+67 0.0 -1.0 -1.0 
+68 0.0 -1.0 -1.0 
+69 0.0 -1.0 -1.0 
+70 1.0 10.0 10.0 
+71 0.0 -1.0 -1.0 
+72 0.0 -1.0 -1.0 
+73 0.0 -1.0 -1.0 
+74 0.0 -1.0 -1.0 
+75 0.0 -1.0 -1.0 
+76 0.0 -1.0 -1.0 
+77 0.0 -1.0 -1.0 
+78 1.0 10.0 10.0 
+79 1.0 10.0 10.0 
+80 0.0 -1.0 -1.0 
+81 0.0 -1.0 -1.0 
+82 0.0 -1.0 -1.0 
+83 0.0 -1.0 -1.0 
+84 0.0 -1.0 -1.0 
+85 0.0 -1.0 -1.0 
+86 0.0 -1.0 -1.0 
+87 0.0 -1.0 -1.0 
+88 1.0 10.0 10.0 
+89 0.0 -1.0 -1.0 
+90 0.0 -1.0 -1.0 
+91 1.0 10.0 10.0 
+92 0.0 -1.0 -1.0 
+93 0.0 -1.0 -1.0 
+94 0.0 -1.0 -1.0 
+95 0.0 -1.0 -1.0 
+96 0.0 -1.0 -1.0 
+97 0.0 -1.0 -1.0 
+98 0.0 -1.0 -1.0 
+99 0.0 -1.0 -1.0 
+100 0.0 -1.0 -1.0 
+101 0.0 -1.0 -1.0 
+102 0.0 -1.0 -1.0 
+103 1.0 10.0 10.0 
+104 0.0 -1.0 -1.0 
+105 0.0 -1.0 -1.0 
+106 0.0 -1.0 -1.0 
+107 0.0 -1.0 -1.0 
+108 0.0 -1.0 -1.0 
+109 0.0 -1.0 -1.0 
+Alfa 
+0 1.0 
+1 1.0 
+2 1.0 
+3 1.0 
+4 1.0 
+5 1.0 
+6 1.0 
+7 1.0 
+8 1.0 
+9 1.0 
+10 1.0 
+11 1.0 
+12 1.0 
+13 1.0 
+14 1.0 
+15 1.0 
+16 1.0 
+17 1.0 
+18 1.0 
+19 1.0 
+20 1.0 
+21 1.0 
+22 1.0 
+23 1.0 
+24 1.0 
+25 1.0 
+26 1.0 
+27 1.0 
+28 1.0 
+29 1.0 
+30 1.0 
+31 1.0 
+32 1.0 
+33 1.0 
+34 1.0 
+35 1.0 
+36 1.0 
+37 1.0 
+38 1.0 
+39 1.0 
+40 1.0 
+41 1.0 
+42 1.0 
+43 1.0 
+44 1.0 
+45 1.0 
+46 1.0 
+47 1.0 
+48 1.0 
+49 1.0 
+50 1.0 
+51 1.0 
+52 1.0 
+53 1.0 
+54 1.0 
+55 1.0 
+56 1.0 
+57 1.0 
+58 1.0 
+59 1.0 
+60 1.0 
+61 1.0 
+62 1.0 
+63 1.0 
+64 1.0 
+65 1.0 
+66 1.0 
+67 1.0 
+68 1.0 
+69 1.0 
+70 1.0 
+71 1.0 
+72 1.0 
+73 1.0 
+74 1.0 
+75 1.0 
+76 1.0 
+77 1.0 
+78 1.0 
+79 1.0 
+80 1.0 
+81 1.0 
+82 1.0 
+83 1.0 
+84 1.0 
+85 1.0 
+86 1.0 
+87 1.0 
+88 1.0 
+89 1.0 
+90 1.0 
+91 1.0 
+92 1.0 
+93 1.0 
+94 1.0 
+95 1.0 
+96 1.0 
+97 1.0 
+98 1.0 
+99 1.0 
+100 1.0 
+101 1.0 
+102 1.0 
+103 1.0 
+104 1.0 
+105 1.0 
+106 1.0 
+107 1.0 
+108 1.0 
+109 1.0
