--- Sim_Settings/set_case9_kinit_1_sp_.txt
@@ -1,12 +1,0 @@
-T_ini: 0.0 
-T_fin: 6000.0 
-Print_steps: 10.0 
-T_max_step: 0.01 
-T_disturb: 2000000.0 
-T_recover: 2000010.0 
-K_ini: 10.5 
-K_fin: 50.0 
-K_step: 0.5 
-Network_file: Networks/case9_kinit_1_sp_.txt 
-Initial_state_file: Initial_States/initstate_case9_kinit_1_sp_.txt 
-Model: sp
